content/media/video/test/test_bug461281.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Media test: Bug 461281 - Ogg Streams with eof marker</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<video id='v'
       src='bug461281.ogg'
       onloadedmetadata='return startTest();'
       onended='return playbackEnded();'></video>
<pre id="test">
<script class="testbody" type="text/javascript">
// Test if the ended event occurs with media without an eos marker
var v = document.getElementById('v');
var endPassed = false;
var completed = false;

function startTest() {
  if (completed)
    return false;

  v.play();
  return false;
}

function playbackEnded() {
  if (completed)
    return false

  completed = true;
  ok(v.ended, "Checking playback has ended");
  SimpleTest.finish();
  return false;
}

SimpleTest.waitForExplicitFinish();
</script>
</pre>
</body>
</html>
